+++ scheduler/indexer/github/helpers.go
@@ -1,0 +1,49 @@
+package githubIndexer
+
+import (
+	"math/rand"
+	"time"
+
+	"github.com/gophr-pm/gophr/lib/model"
+)
+
+var letterRunes = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
+
+// randStringRunes generates a random string n runes long.
+func randStringRunes(n int) string {
+	b := make([]rune, n)
+	for i := range b {
+		b[i] = letterRunes[rand.Intn(len(letterRunes))]
+	}
+	return string(b)
+}
+
+func generateRandomPackageModels(numPackages int) []*models.PackageModel {
+	var pkgModels []*models.PackageModel
+	for i := 0; i < numPackages; i++ {
+		repo := randStringRunes(5)
+		stars := 100
+		exists := true
+		author := randStringRunes(8)
+		versions := []string{randStringRunes(8), randStringRunes(3)}
+		godocURL := randStringRunes(16)
+		indexTime := time.Now()
+		awesomeGo := false
+		searchBlob := randStringRunes(20)
+		description := randStringRunes(40)
+		pkgModel := models.PackageModel{
+			Repo:        &repo,
+			Stars:       &stars,
+			Exists:      &exists,
+			Author:      &author,
+			Versions:    versions,
+			GodocURL:    &godocURL,
+			IndexTime:   &indexTime,
+			AwesomeGo:   &awesomeGo,
+			SearchBlob:  &searchBlob,
+			Description: &description,
+		}
+		pkgModels = append(pkgModels, &pkgModel)
+	}
+	return pkgModels
+}
